// src/API.ts
export type Message = {
  __typename: "Message",
  id: string,
  text?: string | null,
  chatroomID: string,
  userID: string,
  images?: Array< string | null > | null,
  Attachments?: ModelAttachmentConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelAttachmentConnection = {
  __typename: "ModelAttachmentConnection",
  items:  Array<Attachment | null >,
  nextToken?: string | null,
};

export type Attachment = {
  __typename: "Attachment",
  id: string,
  storageKey: string,
  type: AttachmentType,
  width?: number | null,
  height?: number | null,
  duration?: number | null,
  messageID: string,
  chatroomID: string,
  createdAt: string,
  updatedAt: string,
};
// ...
export enum NotificationType {
  APPLIED = "APPLIED",
  COMPLETED = "COMPLETED",
  PAID = "PAID",
}
// ...
export type GetChatRoomQueryVariables = {
  id: string,
};
// ...
export type GetUserChatRoomQueryVariables = {
  id: string,
};